fix(wcb): cast ceil() to int for set_pagination_args + trim PHPStan baseline 75->70
The 5 admin list tables passed ceil() (float) as total_pages where
WP_List_Table::set_pagination_args expects int — a real type error, not a WP
stub limitation. Cast each to (int) ceil(...) and removed the 5 now-resolved
baseline entries. Pays down genuine debt; baseline 75 -> 70. No behavior change
(total_pages was already used as an int by core).

+++ admin/class-admin-candidates.php
@@ -231,7 +231,7 @@ class AdminCandidates extends \WP_List_Table {
 			array(
 				'total_items' => $query->get_total(),
 				'per_page'    => $per_page,
-				'total_pages' => ceil( $query->get_total() / $per_page ),
+				'total_pages' => (int) ceil( $query->get_total() / $per_page ),
 			)
 		);
